// src/app/shared/model/updateCustomerRequest.ts
/**
 * MoHIS Activation API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateCustomerRequest { 
    name?: string;
    description?: string;
    code?: string;
    server?: string;
    numberOfDevices?: number;
}